Move invariant block generation to DAL AST helper.
