+++ workbench/src/Providers/DesignValidator/DesignValidator.js
@@ -1,3 +1,4 @@
+import {getInvariantBlock} from "./AstHelper";
 import {DesignGraph} from "./DesignGraph";
 
 /**
@@ -324,29 +325,6 @@ export class DesignValidator {
      * @param {Object} node
      */
     addInvariants (node) {
-        /**
-         * This entire block is messy because I am constructing
-         * the DAL ast to synthesize the relevant implementation
-         *
-         * TODO:
-         * It will be much cleaner once I add methods in the
-         * dal-ast-js package to generate nodes but I am adding
-         * a todo to remind myself to organize the implementation
-         * below into a more digestable format.
-         *
-         * The structure is:
-         *
-         * - Check if the participants exist
-         * - If they exist
-         *      - get participant values
-         *      - call the invariant
-         *      - if invariant violation
-         *          - log it
-         *
-         * I am rendering the check for participants through the
-         * world state manager. I am also logging the invariant
-         * using the world state manager.
-         */
         if ("body" in node) {
             for (const child of node["body"]) {
                 if (child["type"] === "behavior") {
@@ -357,163 +335,7 @@ export class DesignValidator {
                             continue;
                         }
 
-                        const invCmd = invariant["transformation"].slice(1) +
-                             "_invariant_" + invariant["index"];
-
-                        const invViolationMessage = "f'Semantically invalid state: " + invCmd +
-                             " for " +
-                            invariant["participants"].map((t) => t.value).toString() +
-                            " in transformation " +
-                            invariant["transformation"] + " in behavior " +
-                            invariant["transformBehavior"] + "'";
-
-                        const validParticipants = [];
-                        invariant.participants.forEach((p) => {
-                            if (p.type === "name") {
-                                validParticipants.push(p.value);
-                            }
-                        });
-
-                        // Has participants
-                        const hasParticipants = {
-                            "type": "cmd",
-                            "command": "worldStateManager",
-                            "args": [
-                                {
-                                    "arg": "storeIn",
-                                    "type": "name",
-                                    "value": "hasParticipants",
-                                },
-                                {
-                                    "arg": "cmd",
-                                    "type": "string",
-                                    "value": "hasParticipants",
-                                },
-                                {
-                                    "arg": "participant",
-                                    "type": "list",
-                                    "value": validParticipants,
-                                },
-                            ],
-                        };
-
-                        const callInv = {
-                            "type": "registeredCmd",
-                            "command": "_callIfExist",
-                            "args": [
-                                {
-                                    "arg": null,
-                                    "type": "string",
-                                    "value": "invariantViolated",
-                                },
-                                {
-                                    "arg": null,
-                                    "type": "string",
-                                    "value": invCmd,
-                                },
-                            ].concat(invariant.participants.map((inv) => {
-                                return {
-                                    "arg": null,
-                                    "type": inv.type,
-                                    "value": inv.value,
-                                };
-                            })),
-                        };
-
-                        // Get participant values
-                        const getParticipants = [];
-                        for (const entry of invariant.participants) {
-                            if (entry.type === "name") {
-                                getParticipants.push(
-                                    this.getWorldStateParticipantAST(
-                                        "getValue",
-                                        entry.value,
-                                        entry.value
-                                    )
-                                );
-                            }
-                        }
-
-                        const setInvariantViolation = {
-                            "type": "cmd",
-                            "command": "worldStateManager",
-                            "args": [
-                                {
-                                    "arg": "storeIn",
-                                    "type": "name",
-                                    "value": "hasParticipants",
-                                },
-                                {
-                                    "arg": "cmd",
-                                    "type": "string",
-                                    "value": "setInvariantViolation",
-                                },
-                                {
-                                    "arg": "invariantName",
-                                    "type": "string",
-                                    "value": invCmd,
-                                },
-                                {
-                                    "arg": "invartiantParticipant",
-                                    "type": "string",
-                                    "value": invariant["participants"].map(
-                                        (t) => t.value
-                                    ).toString(),
-                                },
-                                {
-                                    "arg": "protectedBehavior",
-                                    "type": "string",
-                                    "value": invariant.transformBehavior,
-                                },
-                            ],
-                        };
-
-                        // If participants exist
-                        const ifInvariantViolated = {
-                            "type": "if",
-                            "args": [
-                                {
-                                    "arg": null,
-                                    "type": "name",
-                                    "value": "invariantViolated",
-                                },
-                            ],
-                            "body": [
-                                {
-                                    "type": "cmd",
-                                    "command": "display",
-                                    "args": [
-                                        {
-                                            "arg": null,
-                                            "type": "string",
-                                            "value": invViolationMessage,
-                                        },
-                                    ],
-                                },
-                                setInvariantViolation,
-                            ],
-                        };
-
-                        // If participants exist
-                        const ifParticipantsExist = {
-                            "type": "if",
-                            "args": [
-                                {
-                                    "arg": null,
-                                    "type": "name",
-                                    "value": "hasParticipants",
-                                },
-                            ],
-                            "body": [...getParticipants, callInv, ifInvariantViolated],
-                        };
-
-                        const body = [hasParticipants, ifParticipantsExist];
-
-                        const invariantBlock = {
-                            "type": "invariant",
-                            "args": [],
-                            "body": body,
-                        };
+                        const invariantBlock = getInvariantBlock(invariant);
                         child.body.splice(child.body.length - 1, 0, invariantBlock);
                     };
                 } else {
@@ -522,48 +344,6 @@ export class DesignValidator {
             }
         }
     }
-
-    /**
-     * Returns the participant args AST.
-     * @param {String} cmd
-     * @param {String} storeP
-     * @param {String} p
-     * @return {Object}
-     */
-    getWorldStateParticipantAST (cmd, storeP, p) {
-        return {
-            "type": "cmd",
-            "command": "worldStateManager",
-            "args": [
-                {
-                    "arg": "storeIn",
-                    "type": "name",
-                    "value": storeP,
-                },
-                {
-                    "arg": "cmd",
-                    "type": "string",
-                    "value": cmd,
-                },
-                {
-                    "arg": "participant",
-                    "type": "string",
-                    "value": `${p}`,
-                },
-                {
-                    "arg": "type",
-                    "type": "string",
-                    "value": "",
-                },
-                {
-                    "arg": "role",
-                    "type": "string",
-                    "value": "",
-                },
-            ],
-        };
-    }
-
 
     /**
      * Walk path from behavior to behavior and save in path.
